Check exception

+++ AssignmentFPTBook/Controllers/BooksController.cs
@@ -67,14 +67,25 @@ namespace AssignmentFPTBook.Controllers
                 if (image != null && image.ContentLength > 0)
                 {
                     string pic = Path.GetFileName(image.FileName);
+
+                    string extension = Path.GetExtension(image.FileName);
+
                     string path = Path.Combine(Server.MapPath("~/Image/"), pic);
-                    image.SaveAs(path);
 
-                    book.UrlImage = pic;
-                    db.Books.Add(book);
-                    db.SaveChanges();
+                    if (extension.ToLower() == ".jpg" || extension.ToLower() == ".png" || extension.ToLower() == ".jpeg")
+                    {
+                        image.SaveAs(path);
 
-                    return RedirectToAction("Index");
+                        book.UrlImage = pic;
+                        db.Books.Add(book);
+                        db.SaveChanges();
+
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        ViewBag.Extension = "File is invalid. Only accept image file";
+                    }
                 }
                 else
                 {
@@ -123,17 +134,25 @@ namespace AssignmentFPTBook.Controllers
                     string pic = Path.GetFileName(image.FileName);
                     string path = Path.Combine(Server.MapPath("~/Image/"), pic);
                     string oldPath = Request.MapPath(Session["imgPath"].ToString());
-                    image.SaveAs(path);
-
-                    book.UrlImage = pic;
-
-                    db.Entry(book).State = EntityState.Modified;
-                    if (System.IO.File.Exists(oldPath))
+                    string extension = Path.GetExtension(image.FileName);
+                    if (extension.ToLower() == ".jpg" || extension.ToLower() == ".png" || extension.ToLower() == ".jpeg")
                     {
-                        System.IO.File.Delete(oldPath);
+                        image.SaveAs(path);
+
+                        book.UrlImage = pic;
+
+                        db.Entry(book).State = EntityState.Modified;
+                        if (System.IO.File.Exists(oldPath))
+                        {
+                            System.IO.File.Delete(oldPath);
+                        }
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
                     }
-                    db.SaveChanges();
-                    return RedirectToAction("Index");
+                    else
+                    {
+                        ViewBag.Extension = "File is invalid. Only accept image file";
+                    }
                 }
                 else
                 {
